requirements.txt
bs4==0.0.2 
Flask==3.0.3
google-generativeai==0.5.2
langchain-chroma==0.1.0
langchain-community==0.0.34
langchain-core==0.1.45
langchain-google-genai==1.0.2
langchain-text-splitters==0.0.1
numpy==1.26.4
pypdf==4.2.0
python-dotenv==1.0.1
 
 
 